+++ app/db/__init__.py
@@ -1,5 +1,2 @@
-from .card import Card as CardDB
-from .desk import Desk as DeskDB
-from .column import Column as ColumnDB
-from .db import session, Base, engine, database_channel
+from .db import session, Base, engine, database_channel, Desk as DeskDB, Column as ColumnDB, Card as CardDB
